:art: refactor exisitng api calls #17
