+++ lib/github_oauth_handler.js
@@ -16,7 +16,8 @@ module.exports = function oauthHandler (req, reply, tokens, profile) {
     token = JWT.sign(session, process.env.JWT_SECRET);
 
     return reply('<p>Hello ' + profile.name + successMessage + homeLink)
-      .state('token', token);
+      .state('token', token)
+    ;
   }
 
   return reply('Sorry, something went wrong, please try again.').code(401);
